+++ build.clj
@@ -3,7 +3,7 @@
   (:require [clojure.tools.build.api :as b]))
 
 (def lib 'build/micro-x)
-(def version "v0.8.75")
+(def version "v0.11.97")
 (def main 'chat.server)
 (def class-dir "target/classes")
 
@@ -17,12 +17,12 @@
          :ns-compile [main]))
 
 (defn uber [opts]
-  ;; (b/delete {:path "target"})
+  (b/delete {:path "target"})
   (let [opts (uber-opts opts)]
+    (println (str "\nCompiling client ..."))
+    (b/process {:command-args ["clojure" "-M:cljs" "compile" "client"]})
     (println "\nCopying source...")
     (b/copy-dir {:src-dirs ["resources" "src"] :target-dir class-dir})
-    ;; (println (str "\nCompiling client ..."))
-    ;; how to compile cljs?
     (println (str "\nCompiling " main "..."))
     (b/compile-clj opts)
     (println "\nBuilding JAR...")