+++ src/chat/server.clj
@@ -5,6 +5,7 @@
             [clojure.java.io :as io]
             [clojure.string :as str]
             [hato.client :as hc]
+            [muuntaja.middleware :as mw]
             [reitit.ring :as rr]
             [ring.adapter.jetty :as adapter]
             [ring.util.anti-forgery :refer [anti-forgery-field]]
@@ -15,9 +16,16 @@
             [ring.websocket.keepalive :as wska]
             [taoensso.telemere :as t]))
 
-(def ^:private version "v0.8.75")
+(def debug? (System/getenv "MX3_DEV"))
 
-(def ^:pricate url "https://l22.melt.kyutech.ac.jp/api/user/")
+(t/set-min-level! (if debug? :debug :info))
+
+(def ^:private version "v0.11.97")
+
+(def ^:private l22
+  (if debug?
+    "http://localhost:3090/"
+    "https://l22.melt.kyutech.ac.jp/"))
 
 (defn make-chat-handler []
   (let [writer  (a/chan)
@@ -32,7 +40,7 @@
     (-> (resp/response
          (str
           "<!DOCTYPE html><title>MX3</title>
-           <h1>Micro X for hkimura classes</h1>
+           <h1>Micro X version3</h1>
            <body style='font-family:sans-serif;'>
            <form method='post'>"
           (anti-forgery-field)
@@ -48,11 +56,12 @@
         (resp/charset "UTF-8"))))
 
 (defn login! [{{:keys [login password]} :params}]
-  (if (System/getenv "MX3_DEBUG")
+  (if debug?
     (-> (resp/redirect "/index")
         (assoc-in [:session :identity] login))
     (try
-      (let [resp (hc/get (str url login) {:timeout 3000 :as :json})]
+      (let [resp (hc/get (str l22 "api/user/"login)
+                         {:timeout 3000 :as :json})]
         (if (and (some? resp)
                  (hashers/check password (get-in resp [:body :password])))
           (-> (resp/redirect "/index")
@@ -74,11 +83,42 @@
           (resp/content-type "text/html")
           (resp/charset "UTF-8")))))
 
+;; must be rewritten with java-time. agry.
+(defn- utime [t]
+  (cond
+    debug? "1"
+    (< (+ (* 8 60) 50) t  (+ (* 10 60) 20)) "1"
+    (< (+ (* 10 60) 30) t (+ (* 12 60))) "2"
+    :else "0"))
+
+(defn- uhour []
+  (let [[wd _ _ hhmmss] (-> (str (java.util.Date.))
+                            (str/split #"\s"))
+        [hh mm] (str/split hhmmss #":")
+        t (+ (* 60 (Long/parseLong hh)) (Long/parseLong mm))]
+    (str/lower-case (str
+                     (if debug? "wed" wd)
+                     (utime t)))))
+
+(defn user-random [_]
+  (-> (hc/get (str l22 "api/user/" (uhour) "/randomly")
+              {:as :json :timeout 1000})
+      :body))
+
+(comment
+  (user-random nil)
+  :rcf)
+
 (defn make-app-handler []
   (rr/ring-handler
    (rr/router [["/chat" {:middleware [[wst/wrap-websocket-transit]
                                       [wska/wrap-websocket-keepalive]]}
                 ["" (make-chat-handler)]]
+               ["/api" {:middleware [[def/wrap-defaults def/api-defaults]
+                                     mw/wrap-format
+                                     mw/wrap-params]}
+                ["/user-random" {:get (fn [_]
+                                 (resp/response (user-random nil)))}]]
                ["" {:middleware [[def/wrap-defaults def/site-defaults]]}
                 ["/" {:get login :post login!}]
                 ["/logout" (fn [_]
@@ -95,10 +135,14 @@
 
 (def server (atom nil))
 
-(defn start []
-  (when-not (some? @server)
-    (reset! server (run-server {:port 8080 :join? false}))
-    (println "server started in port 8080.")))
+(defn start
+  ([] (if-let [p (System/getenv "PORT")]
+        (start {:port (Long/parseLong p)})
+        (start {:port 8080})))
+  ([{:keys [port]}]
+   (when-not (some? @server)
+     (reset! server (run-server {:port port :join? false}))
+     (println "server started in port " port "."))))
 
 (defn stop []
   (when (some? @server)