+++ systemd-service/start.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+PORT=8080 java -jar micro-x.jar